Incorporated large structure instead of the naive basic structure
